adding components for scunt & fixing style
